requirements.txt
fastapi
uvicorn
streamlit
google-cloud-texttospeech
pydub
pyaudio
requests